+++ packages/react-webview/src/components/Badge/index.tsx
@@ -1,19 +1,26 @@
-
-import React from 'react';
-import { BadgeContainer } from './styles';
+import React from 'react'
+import { BadgeContainer } from './styles'
 
 export interface BadgeProps {
-  backgroundColor?: string
-  color?: string 
+  color?: string
   children: React.ReactNode
+  type?: 'primary' | 'secondary'
 }
 
-export function Badge({backgroundColor = '$green-600', children, color = '$white'}: BadgeProps) {
-  return (
-    <BadgeContainer css={{backgroundColor, color}}>
-      {children}
-    </BadgeContainer>
-  )
+export function Badge({
+  color = '$green-600',
+  children,
+  type = 'primary',
+}: BadgeProps) {
+  const isPrimary = type === 'primary'
+  const dynamicStyles = {
+    backgroundColor: isPrimary ? `${color}` : 'transparent',
+    color: isPrimary ? '$white' : `${color}`,
+    ...(type === 'secondary' && { border: `1.5px solid ${color}` }),
+  }
+
+  console.log(dynamicStyles)
+  return <BadgeContainer css={dynamicStyles}>{children}</BadgeContainer>
 }
 
 Badge.displayName = 'Badge'