+++ packages/react-webview/src/components/Badge/styles.ts
@@ -1,11 +1,10 @@
-
-import { styled } from '../../styles';
+import { styled } from '../../styles'
 
 export const BadgeContainer = styled('div', {
-  borderRadius: '16px',
-  padding: '0.1rem 0.3rem',
-  fontWeight: 500,
-  fontSize: '0.625rem',
+  borderRadius: '$lg',
+  padding: '2px 8px',
+  fontWeight: '$medium',
+  fontSize: '$sm',
   textAlign: 'center',
-  display: 'inline-block'
+  display: 'inline-block',
 })