+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @allan-koin/eslint-config
 
+## 2.2.0
+
+### Minor Changes
+
+- add secondary type in badge component
+
 ## 2.1.0
 
 ### Minor Changes